refactor: index 적용
