+++ frontend/src/utils/index.ts
@@ -2,4 +2,7 @@ export * from './date';
 export { default as isExistentElement } from './isExistentElement';
 export { default as scrollToTop } from './scrollToTop';
 export { default as debounce } from './debounce';
+export { default as hasFinalConsonant } from './hasFinalConsonant';
+export { default as substituteString } from './substituteString';
+export { default as calculateParticle } from './calculateParticle';
 export * from './media';
